separate network availability and offline service viewmodels
