+++ Core/Core/NetworkAvailability/NetworkAvailabilityService.swift
@@ -88,22 +88,3 @@ public final class NetworkAvailabilityServiceLive: NetworkAvailabilityService {
         statusSubject
     }
 }
-
-class NetworkAvailabilityObservableModel: ObservableObject {
-    @Published var networkAvailabilityStatus: NetworkAvailabilityStatus = .disconnected
-
-    private let networkAvailabilityService = NetworkAvailabilityServiceLive()
-    private var subscriptions = Set<AnyCancellable>()
-
-    init() {
-        unowned let unownedSelf = self
-
-        networkAvailabilityService.startMonitoring()
-
-        networkAvailabilityService
-            .startObservingStatus()
-            .receive(on: DispatchQueue.main)
-            .sink { unownedSelf.networkAvailabilityStatus = $0 }
-            .store(in: &subscriptions)
-    }
-}
